+++ src/settings.ts
@@ -91,6 +91,16 @@ export interface SettingsIOS {
    */
   barCollapsingEnabled?: boolean;
 
+  /**
+   * A value that specifies whether Safari should enter Reader mode, if it is available.
+   *
+   * **Available on**: iOS >= 11.0.
+   *
+   * **Note**: if the value is invalid or if the current iOS version
+   * is < 11.0, this setting will be ignored.
+   */
+  entersReaderIfAvailable?: boolean;
+
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   [key: string]: any;
 }
@@ -153,6 +163,9 @@ function sanitizeIOS(settings?: SettingsIOS): SettingsIOS {
     sanitized.barCollapsingEnabled = settings.barCollapsingEnabled;
   }
 
+  if (typeof settings.entersReaderIfAvailable === 'boolean') {
+    sanitized.entersReaderIfAvailable = settings.entersReaderIfAvailable;
+  }
   return sanitized;
 }