+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Bump Kotlin to version `1.3.72`.
+
 ## [3.2.0] - 2020-02-29
 
 ### Added
@@ -199,6 +203,7 @@ The following functions have been deprecated. Please see
 
 - `openInApp`: open a valid http(s) URL with an in-app browser.
 
+[Unreleased]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.2.0...HEAD
 [3.2.0]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.1.4...v3.2.0
 [3.1.4]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.1.3...v3.1.4
 [3.1.3]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.1.2...v3.1.3