+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Bump Kotlin to version `1.3.72`.
+- Bump most JS/TS dependencies.
 
 ## [3.2.0] - 2020-02-29