+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.2.1] - 2020-05-17
+
 ### Changed
 - Bump Kotlin to version `1.3.72`.
 - Bump most JS/TS dependencies.
@@ -204,7 +207,8 @@ The following functions have been deprecated. Please see
 
 - `openInApp`: open a valid http(s) URL with an in-app browser.
 
-[Unreleased]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.2.0...HEAD
+[Unreleased]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.2.1...HEAD
+[3.2.1]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.1.4...v3.2.0
 [3.1.4]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.1.3...v3.1.4
 [3.1.3]: https://github.com/matei-radu/react-native-in-app-browser/compare/v3.1.2...v3.1.3